third_party/blink/web_tests/external/wpt/svg/animations/animate-stop-currentcolor-escaped-expected.txt

This is a testharness.js-based test.
[FAIL] Animating 'stop-color' from 'currentc\\olor'
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
